move some logic to repository layer
