+++ src/habit/habit.service.ts
@@ -1,8 +1,6 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { HabitRepository } from './habit.repository';
 import { CreateHabitPayload } from './payload/create.habit.payload';
-import { CreateHabitType } from './type/create.habit.type';
-import { HabitRecordDayConst } from './const/habitRecordDay.const';
 
 @Injectable()
 export class HabitService {
@@ -12,8 +10,7 @@ export class HabitService {
     userId: string,
     payload: CreateHabitPayload,
   ): Promise<void> {
-    const { title, action, value, unit, time, startDate, endDate, days } =
-      payload;
+    const { startDate, endDate } = payload;
 
     // startDate 검증
     const date = new Date();
@@ -26,29 +23,6 @@ export class HabitService {
       throw new BadRequestException('invalid endDate');
     }
 
-    // 요일을 bit 합으로 변환
-    const dayBit = days.reduce((acc, cur) => acc + HabitRecordDayConst[cur], 0);
-
-    // time 을 Date 타입으로 변환
-    let habitTime = null;
-    if (time) {
-      const [hh, mm] = time.split(':');
-      habitTime = date;
-      habitTime.setUTCHours(+hh, +mm, 0, 0);
-    }
-
-    const habitData: CreateHabitType = {
-      userId,
-      title,
-      action,
-      value,
-      unit,
-      time: habitTime,
-      startDate: new Date(startDate),
-      endDate: endDate ? new Date(endDate) : null,
-      days: dayBit,
-    };
-
-    await this.habitRepository.create(habitData);
+    await this.habitRepository.create(userId, payload);
   }
 }
